api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
numpy==1.23.5
pandas==1.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
